retrieving process info from linux procfs; automated tests for the process launching code

+++ src/process.cpp
@@ -33,7 +33,8 @@ sdb::stop_reason::stop_reason(int wait_status) {
     }
 }
 
-std::unique_ptr<sdb::process> sdb::process::launch(const std::filesystem::path& path) {
+std::unique_ptr<sdb::process> sdb::process::launch(
+    const std::filesystem::path& path, bool debug) {
     // Pipe for communicating errors when spawning new processes
     pipe channel(true);
 
@@ -45,7 +46,7 @@ std::unique_ptr<sdb::process> sdb::process::launch(const std::filesystem::path& 
     // In child process
     if (pid == 0) {
         channel.close_read();
-        if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+        if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
             exit_with_perror(channel, "Tracing failed");
         }
 
@@ -64,8 +65,11 @@ std::unique_ptr<sdb::process> sdb::process::launch(const std::filesystem::path& 
         error::send(std::string(chars, chars + data.size()));
     }
 
-    std::unique_ptr<process> proc (new process(pid, true));
-    proc->wait_on_signal();
+    std::unique_ptr<process> proc (new process(pid, true, debug));
+
+    if (debug) {
+        proc->wait_on_signal();
+    }
 
     return proc;
 }
@@ -79,7 +83,7 @@ std::unique_ptr<sdb::process> sdb::process::attach(pid_t pid) {
         error::send_errno("Could not attach");
     }
 
-    std::unique_ptr<process> proc (new process(pid, false));
+    std::unique_ptr<process> proc (new process(pid, false, true));
     proc->wait_on_signal();
 
     return proc;
@@ -99,12 +103,14 @@ sdb::process::~process() {
     if (pid_ != 0) {
         int status;
         // Inferior process must be stopped for detach to work
-        if (state_ == process_state::running) {
-            kill(pid_, SIGSTOP);
-            waitpid(pid_, &status, 0);
+        if (is_attached_) {
+            if (state_ == process_state::running) {
+                kill(pid_, SIGSTOP);
+                waitpid(pid_, &status, 0);
+            }
+            ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+            kill(pid_, SIGCONT);
         }
-        ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-        kill(pid_, SIGCONT);
 
         if (terminate_on_end_) {
             kill(pid_, SIGKILL);
